+++ tests/helpers/fixtures/fake_filter_v2.ts
@@ -1,5 +1,7 @@
+import { calculateChecksum } from '../../../tools/resources/download-filters';
+
 const currentDateWithoutMS = new Date().toISOString().slice(0, -5);
-const fakeFilterV2 = `! Title: Fake filter
+const fakeFilterWithoutChecksum = `! Title: Fake filter
 ! Description: Fake filter
 ! Version: 2.0.0.0
 ! TimeUpdated: ${currentDateWithoutMS}+00:00
@@ -8,4 +10,6 @@ const fakeFilterV2 = `! Title: Fake filter
 ||example.com^$document
 ||example.org^##h1`;
 
-export { fakeFilterV2 };
+const checksum = calculateChecksum(fakeFilterWithoutChecksum);
+export const fakeFilterV2 = `! Checksum: ${checksum}
+${fakeFilterWithoutChecksum}`;